Move ApplicationWorksheetFunctionInspectionTests QuickFix tests into separate files
